move service tests to dedicated folder to mirror the source code architecture
