+++ tests/base.py
@@ -1,4 +1,3 @@
-import abc
 import contextlib
 import os.path
 import shutil
@@ -8,10 +7,9 @@ import unittest
 import unittest.mock
 
 import pytest
-import responses
 
 from bugwarrior import config, services
-from bugwarrior.config import schema, validation
+from bugwarrior.config import validation
 from bugwarrior.config.load import format_config
 
 
@@ -101,27 +99,6 @@ def register_services(mapping=None):
         yield mapping
 
 
-class AbstractServiceTest(abc.ABC):
-    """Ensures that certain test methods are implemented for each service."""
-
-    @abc.abstractmethod
-    def test_to_taskwarrior(self):
-        """Test Service.to_taskwarrior()."""
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def test_issues(self):
-        """
-        Test Service.issues().
-
-        - When the API is accessed via requests, use the responses library to
-        mock requests.
-        - When the API is accessed via a third party library, substitute a fake
-        implementation class for it.
-        """
-        raise NotImplementedError
-
-
 class ConfigTest(unittest.TestCase):
     """
     Creates config files, configures the environment, and cleans up afterwards.
@@ -182,39 +159,3 @@ class ConfigTest(unittest.TestCase):
 
         # We may want to use this assertion more than once per test.
         self.caplog.clear()
-
-
-class ServiceTest(ConfigTest):
-    GENERAL_CONFIG = {'annotation_length': 100, 'description_length': 100}
-    SERVICE_CONFIG = {}
-
-    @classmethod
-    def setUpClass(cls):
-        cls.maxDiff = None
-
-    def get_mock_service(
-        self,
-        service_class,
-        section='unspecified',
-        config_overrides=None,
-        general_overrides=None,
-    ):
-        options = {
-            'general': {**self.GENERAL_CONFIG, 'targets': [section]},
-            section: {**self.SERVICE_CONFIG.copy(), 'target': section},
-        }
-        if config_overrides:
-            options[section].update(config_overrides)
-        if general_overrides:
-            options['general'].update(general_overrides)
-
-        service_config = service_class.CONFIG_SCHEMA(**options[section])
-        main_config = schema.MainSectionConfig(**options['general'])
-
-        return service_class(service_config, main_config)
-
-    @staticmethod
-    def add_response(url, method='GET', **kwargs):
-        responses.add(
-            responses.Response(url=url, method=method, match_querystring=True, **kwargs)
-        )
