+++ tests/services/test_bitbucket.py
